refactor: adjust helpers structure
